+++ mirrord/agent/src/dns.rs
@@ -11,7 +11,7 @@ use tokio::sync::{mpsc::Receiver, oneshot::Sender};
 use tracing::{error, trace};
 use trust_dns_resolver::{system_conf::parse_resolv_conf, AsyncResolver, Hosts};
 
-use crate::{error::Result, runtime::set_namespace};
+use crate::error::Result;
 
 #[derive(Debug)]
 pub struct DnsRequest {
@@ -59,13 +59,6 @@ async fn dns_lookup(root_path: &Path, host: String) -> RemoteResult<DnsLookup> {
 /// Reads a `DnsRequest` from `rx` and returns the resolved addresses (or `Error`) through
 /// `DnsRequest::tx`.
 pub async fn dns_worker(mut rx: Receiver<DnsRequest>, pid: Option<u64>) -> Result<()> {
-    if let Some(pid) = pid {
-        let namespace = PathBuf::from("/proc")
-            .join(PathBuf::from(pid.to_string()))
-            .join(PathBuf::from("ns"));
-        set_namespace(namespace.join(PathBuf::from("net")))?;
-    }
-
     let root_path = pid
         .map(|pid| PathBuf::from("/proc").join(pid.to_string()).join("root"))
         .unwrap_or_else(|| PathBuf::from("/"));