+++ mirrord/config/src/network.rs
@@ -3,7 +3,7 @@ use schemars::JsonSchema;
 
 use crate::{
     config::{from_env::FromEnv, source::MirrordConfigSource, ConfigError},
-    incoming::IncomingConfig,
+    incoming::{IncomingConfig, IncomingFileConfig},
     outgoing::{OutgoingConfig, OutgoingFileConfig},
     util::MirrordToggleableConfig,
 };
@@ -31,11 +31,8 @@ use crate::{
 #[config(map_to = "NetworkFileConfig", derive = "JsonSchema")]
 #[cfg_attr(test, config(derive = "PartialEq, Eq"))]
 pub struct NetworkConfig {
-    /// Mode of operation for incoming network requests in mirrord, supports `mirror` or `steal`:
-    ///
-    /// - `mirror`: mirror incoming requests to the remote pod to the local process;
-    /// - `steal`: redirect incoming requests to the remote pod to the local process
-    #[config(env = "MIRRORD_AGENT_TCP_STEAL_TRAFFIC", default)]
+    /// Handles incoming network traffic, see [`IncomingConfig`] for more details.
+    #[config(toggleable, nested)]
     pub incoming: IncomingConfig,
 
     /// Tunnel outgoing network operations through mirrord.
@@ -49,16 +46,13 @@ pub struct NetworkConfig {
 
 impl MirrordToggleableConfig for NetworkFileConfig {
     fn disabled_config() -> Result<Self::Generated, ConfigError> {
-        let incoming = FromEnv::new("MIRRORD_AGENT_TCP_STEAL_TRAFFIC")
-            .source_value()
-            .transpose()?
-            .unwrap_or(IncomingConfig::Mirror);
         let dns = FromEnv::new("MIRRORD_REMOTE_DNS")
             .source_value()
             .transpose()?
             .unwrap_or(false);
+
         Ok(NetworkConfig {
-            incoming,
+            incoming: IncomingFileConfig::disabled_config()?,
             dns,
             outgoing: OutgoingFileConfig::disabled_config()?,
         })
@@ -70,11 +64,28 @@ mod tests {
     use rstest::rstest;
 
     use super::*;
-    use crate::{config::MirrordConfig, util::testing::with_env_vars};
+    use crate::{
+        config::MirrordConfig,
+        incoming::{IncomingConfig, IncomingMode},
+        util::testing::with_env_vars,
+    };
 
     #[rstest]
     fn default(
-        #[values((None, IncomingConfig::Mirror), (Some("false"), IncomingConfig::Mirror), (Some("true"), IncomingConfig::Steal))]
+        #[values(
+            (
+                None,
+                IncomingConfig { mode: IncomingMode::Mirror, http_header_filter: Default::default(), }
+            ),
+            (
+                Some("false"),
+                IncomingConfig { mode: IncomingMode::Mirror, http_header_filter: Default::default(), }
+            ),
+            (
+                Some("true"),
+                IncomingConfig { mode: IncomingMode::Steal, http_header_filter: Default::default(), }
+            ),
+        )]
         incoming: (Option<&str>, IncomingConfig),
         #[values((None, true), (Some("false"), false))] dns: (Option<&str>, bool),
     ) {